chore(kitchenpos): 애그리거트 기준으로 도메인 패키지 나누기
