+++ src/main/java/kitchenpos/menu/domain/menu/JpaMenuRepository.java
@@ -1,7 +1,5 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.menu;
 
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
