+++ src/main/java/kitchenpos/menu/domain/menu/MenuProduct.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.menu.domain.menu;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,13 +11,12 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
 
-import java.math.BigDecimal;
 import java.util.UUID;
-import kitchenpos.menu.domain.Menu;
+import kitchenpos.product.domain.product.Product;
 
-@Table(name = "order_line_item")
+@Table(name = "menu_product")
 @Entity
-public class OrderLineItem {
+public class MenuProduct {
     @Column(name = "seq")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
@@ -25,22 +24,19 @@ public class OrderLineItem {
 
     @ManyToOne(optional = false)
     @JoinColumn(
-        name = "menu_id",
+        name = "product_id",
         columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_order_line_item_to_menu")
+        foreignKey = @ForeignKey(name = "fk_menu_product_to_product")
     )
-    private Menu menu;
+    private Product product;
 
     @Column(name = "quantity", nullable = false)
     private long quantity;
 
     @Transient
-    private UUID menuId;
+    private UUID productId;
 
-    @Transient
-    private BigDecimal price;
-
-    public OrderLineItem() {
+    public MenuProduct() {
     }
 
     public Long getSeq() {
@@ -51,12 +47,12 @@ public class OrderLineItem {
         this.seq = seq;
     }
 
-    public Menu getMenu() {
-        return menu;
+    public Product getProduct() {
+        return product;
     }
 
-    public void setMenu(final Menu menu) {
-        this.menu = menu;
+    public void setProduct(final Product product) {
+        this.product = product;
     }
 
     public long getQuantity() {
@@ -67,19 +63,11 @@ public class OrderLineItem {
         this.quantity = quantity;
     }
 
-    public UUID getMenuId() {
-        return menuId;
+    public UUID getProductId() {
+        return productId;
     }
 
-    public void setMenuId(final UUID menuId) {
-        this.menuId = menuId;
-    }
-
-    public BigDecimal getPrice() {
-        return price;
-    }
-
-    public void setPrice(final BigDecimal price) {
-        this.price = price;
+    public void setProductId(final UUID productId) {
+        this.productId = productId;
     }
 }
