+++ src/main/java/kitchenpos/menu/domain/menugroup/JpaMenuGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.domain.menugroup;
 
 import org.springframework.data.jpa.repository.JpaRepository;
