+++ src/main/java/kitchenpos/order/domain/order/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.domain;
+package kitchenpos.order.domain.order;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,12 +11,13 @@ import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.persistence.Transient;
 
+import java.math.BigDecimal;
 import java.util.UUID;
-import kitchenpos.product.domain.Product;
+import kitchenpos.menu.domain.menu.Menu;
 
-@Table(name = "menu_product")
+@Table(name = "order_line_item")
 @Entity
-public class MenuProduct {
+public class OrderLineItem {
     @Column(name = "seq")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
@@ -24,19 +25,22 @@ public class MenuProduct {
 
     @ManyToOne(optional = false)
     @JoinColumn(
-        name = "product_id",
+        name = "menu_id",
         columnDefinition = "binary(16)",
-        foreignKey = @ForeignKey(name = "fk_menu_product_to_product")
+        foreignKey = @ForeignKey(name = "fk_order_line_item_to_menu")
     )
-    private Product product;
+    private Menu menu;
 
     @Column(name = "quantity", nullable = false)
     private long quantity;
 
     @Transient
-    private UUID productId;
+    private UUID menuId;
 
-    public MenuProduct() {
+    @Transient
+    private BigDecimal price;
+
+    public OrderLineItem() {
     }
 
     public Long getSeq() {
@@ -47,12 +51,12 @@ public class MenuProduct {
         this.seq = seq;
     }
 
-    public Product getProduct() {
-        return product;
+    public Menu getMenu() {
+        return menu;
     }
 
-    public void setProduct(final Product product) {
-        this.product = product;
+    public void setMenu(final Menu menu) {
+        this.menu = menu;
     }
 
     public long getQuantity() {
@@ -63,11 +67,19 @@ public class MenuProduct {
         this.quantity = quantity;
     }
 
-    public UUID getProductId() {
-        return productId;
+    public UUID getMenuId() {
+        return menuId;
     }
 
-    public void setProductId(final UUID productId) {
-        this.productId = productId;
+    public void setMenuId(final UUID menuId) {
+        this.menuId = menuId;
+    }
+
+    public BigDecimal getPrice() {
+        return price;
+    }
+
+    public void setPrice(final BigDecimal price) {
+        this.price = price;
     }
 }
