+++ src/main/java/kitchenpos/order/domain/order/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.order;
 
 public enum OrderStatus {
     WAITING, ACCEPTED, SERVED, DELIVERING, DELIVERED, COMPLETED
