+++ src/main/java/kitchenpos/order/domain/order/OrderType.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.order;
 
 public enum OrderType {
     DELIVERY, TAKEOUT, EAT_IN
