+++ src/main/java/kitchenpos/order/domain/ordertable/JpaOrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.ordertable;
 
 import org.springframework.data.jpa.repository.JpaRepository;
