+++ src/main/java/kitchenpos/order/domain/ordertable/OrderTableRepository.java
@@ -1,8 +1,9 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.domain.ordertable;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import kitchenpos.order.domain.ordertable.OrderTable;
 
 public interface OrderTableRepository {
     OrderTable save(OrderTable orderTable);
