+++ src/main/java/kitchenpos/product/domain/product/JpaProductRepository.java
@@ -1,7 +1,5 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.domain.product;
 
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
