+++ src/main/java/kitchenpos/product/domain/product/ProductRepository.java
@@ -1,9 +1,8 @@
-package kitchenpos.product.domain;
+package kitchenpos.product.domain.product;
 
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
-import kitchenpos.product.domain.Product;
 
 public interface ProductRepository {
     Product save(Product product);
