+++ src/test/java/kitchenpos/application/menu/application/MenuGroupServiceTest.java
@@ -2,8 +2,8 @@ package kitchenpos.application.menu.application;
 
 import kitchenpos.application.fakeobject.InMemoryMenuGroupRepository;
 import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menu.domain.menugroup.MenuGroup;
+import kitchenpos.menu.domain.menugroup.MenuGroupRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
