+++ src/test/java/kitchenpos/application/menu/application/MenuServiceTest.java
@@ -5,12 +5,12 @@ import kitchenpos.application.fakeobject.InMemoryMenuGroupRepository;
 import kitchenpos.application.fakeobject.InMemoryMenuRepository;
 import kitchenpos.application.fakeobject.InMemoryProductRepository;
 import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menugroup.MenuGroupRepository;
+import kitchenpos.menu.domain.menu.MenuProduct;
+import kitchenpos.menu.domain.menu.MenuRepository;
+import kitchenpos.product.domain.product.Product;
+import kitchenpos.product.domain.product.ProductRepository;
 import kitchenpos.common.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
