+++ src/test/java/kitchenpos/application/order/application/OrderServiceTest.java
@@ -4,14 +4,14 @@ import kitchenpos.application.fakeobject.FakeKitchenridersClient;
 import kitchenpos.application.fakeobject.InMemoryMenuRepository;
 import kitchenpos.application.fakeobject.InMemoryOrderRepository;
 import kitchenpos.application.fakeobject.InMemoryOrderTableRepository;
-import kitchenpos.menu.domain.MenuRepository;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderLineItem;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.OrderTableRepository;
-import kitchenpos.order.domain.OrderType;
+import kitchenpos.menu.domain.menu.MenuRepository;
+import kitchenpos.order.domain.order.Order;
+import kitchenpos.order.domain.order.OrderLineItem;
+import kitchenpos.order.domain.order.OrderRepository;
+import kitchenpos.order.domain.order.OrderStatus;
+import kitchenpos.order.domain.ordertable.OrderTable;
+import kitchenpos.order.domain.ordertable.OrderTableRepository;
+import kitchenpos.order.domain.order.OrderType;
 import kitchenpos.order.application.OrderService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
