+++ src/test/java/kitchenpos/application/product/application/ProductServiceTest.java
@@ -3,11 +3,11 @@ package kitchenpos.application.product.application;
 import kitchenpos.application.fakeobject.FakePurgomalumClient;
 import kitchenpos.application.fakeobject.InMemoryMenuRepository;
 import kitchenpos.application.fakeobject.InMemoryProductRepository;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.menu.domain.menu.Menu;
+import kitchenpos.menu.domain.menu.MenuRepository;
 import kitchenpos.product.application.ProductService;
-import kitchenpos.product.domain.Product;
-import kitchenpos.product.domain.ProductRepository;
+import kitchenpos.product.domain.product.Product;
+import kitchenpos.product.domain.product.ProductRepository;
 import kitchenpos.common.infra.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
